+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 Extends support of the SDK to OneView REST API version 1600 (OneView v5.20).
 
 #### Features supported with the current release
+- Appliance SNMPv1 Trap Destinations
 - Certificates Server
 - Enclosures
+- Enclosure Groups
 - Ethernet network
 - FC network
 - FCOE network
@@ -12,6 +14,7 @@ Extends support of the SDK to OneView REST API version 1600 (OneView v5.20).
 - Hypervisor Managers
 - Interconnects
 - Interconnect Types
+- Logical Enclosures
 - Logical Interconnects
 - Logical Interconnect Groups
 - Network set